DoAnalysisOnClassPath.java: added cleanup code on the csv-dumper facade.
ASTNodesAtRangeFinder.java, ASTNodeUnitBridge.java: Changed the parameter on the constructor to a more general one.

FeatureModelInstrumentorTransformer.java: added some benchmarks on the the unit-to-astnode part of the code. It was found that there is a huge bottleneck there. A new approach must be taken.

SetUtil.java: fixed #configurationSet() not working properly since some code was refactored (the way Features were passed around, as Strings or IFeatures).

WriterFacadeForAnalysingMM.java: added some comments about how this class should be extended to a more general functionality.



git-svn-id: https://subversion.assembla.com/svn/cide_ei_rmk@46 6c4dff51-b634-49cf-a5fd-882218f1ddba