Transformed the A1 estimation from
the sum of the time to analyze all configurations whose body size > 0

to

average between the time to analyze the largest and the smallest body. As SOOT can't analyze a body with no units in it, if no body with size > 0 and < largest is found, than the smallest body is to be the original method, and the average is apllied.

git-svn-id: https://subversion.assembla.com/svn/cide_ei_rmk@137 6c4dff51-b634-49cf-a5fd-882218f1ddba